Fixed System Tests
Some of the tests were broken because they were checking for values before they were loaded so it would fail. I fixed this by putting a wait in. The User Registration test also fails if the user exists in the database as it does not allow for duplicate accounts so it passes if the test user is deleted from the database.